+++ app/gui/src/project-view/components/GraphEditor/upload.ts
@@ -1,15 +1,14 @@
 import { Awareness } from '@/stores/awareness'
+import { ProjectFiles, useProjectFiles } from '@/stores/projectFiles'
 import { Vec2 } from '@/util/data/vec2'
 import type { DataServer } from '@/util/net/dataServer'
 import { Keccak, sha3_224 as SHA3 } from '@noble/hashes/sha3'
 import type { Hash } from '@noble/hashes/utils'
-import { bytesToHex } from '@noble/hashes/utils'
 import { markRaw, toRaw } from 'vue'
 import { escapeTextLiteral } from 'ydoc-shared/ast/text'
 import type { LanguageServer } from 'ydoc-shared/languageServer'
-import { ErrorCode, RemoteRpcError } from 'ydoc-shared/languageServer'
 import type { Path, StackItem, Uuid } from 'ydoc-shared/languageServerTypes'
-import { Err, Ok, withContext, type Result } from 'ydoc-shared/util/data/result'
+import { Err, Ok, type Result } from 'ydoc-shared/util/data/result'
 
 // === Constants ===
 
@@ -47,13 +46,17 @@ export class Uploader {
   private checksum: Hash<Keccak>
   private uploadedBytes: bigint
   private stackItem: StackItem
+  private awareness: Awareness
+  private projectFiles: ProjectFiles
 
   private constructor(
-    private rpc: LanguageServer,
-    private binary: DataServer,
-    private awareness: Awareness,
+    projectStore: {
+      projectRootId: Promise<Uuid | undefined>
+      lsRpcConnection: LanguageServer
+      dataConnection: DataServer
+      awareness: Awareness
+    },
     private file: File,
-    private projectRootId: Uuid,
     private position: Vec2,
     private isOnLocalBackend: boolean,
     private disableDirectRead: boolean,
@@ -62,14 +65,18 @@ export class Uploader {
     this.checksum = SHA3.create()
     this.uploadedBytes = BigInt(0)
     this.stackItem = markRaw(toRaw(stackItem))
+    this.awareness = projectStore.awareness
+    this.projectFiles = useProjectFiles(projectStore)
   }
 
   /** Constructor */
   static Create(
-    rpc: LanguageServer,
-    binary: DataServer,
-    projectRootId: Uuid,
-    awareness: Awareness,
+    projectStore: {
+      projectRootId: Promise<Uuid | undefined>
+      lsRpcConnection: LanguageServer
+      dataConnection: DataServer
+      awareness: Awareness
+    },
     file: File,
     position: Vec2,
     isOnLocalBackend: boolean,
@@ -77,11 +84,8 @@ export class Uploader {
     stackItem: StackItem,
   ): Uploader {
     return new Uploader(
-      rpc,
-      binary,
-      awareness,
+      projectStore,
       file,
-      projectRootId,
       position,
       isOnLocalBackend,
       disableDirectRead,
@@ -100,20 +104,29 @@ export class Uploader {
     ) {
       return Ok({ source: 'FileSystemRoot', name: this.file.path })
     }
-    const dataDirExists = await this.ensureDataDirExists()
+    const rootId = await this.projectFiles.projectRootId
+    if (rootId == null) return Err('Could not identify project root.')
+    const dataDirPath = { rootId, segments: [DATA_DIR_NAME] }
+    const dataDirExists = await this.projectFiles.ensureDirExists(dataDirPath)
     if (!dataDirExists.ok) return dataDirExists
-    const name = await this.pickUniqueName(this.file.name)
+    const name = await this.projectFiles.pickUniqueName(dataDirPath, this.file.name)
     if (!name.ok) return name
     this.awareness.addOrUpdateUpload(name.value, {
       sizePercentage: 0,
       position: this.position,
       stackItem: this.stackItem,
     })
-    const remotePath: Path = { rootId: this.projectRootId, segments: [DATA_DIR_NAME, name.value] }
+    const remotePath: Path = { rootId, segments: [DATA_DIR_NAME, name.value] }
     const cleanup = this.cleanup.bind(this, name.value)
     const writableStream = new WritableStream<Uint8Array>({
       write: async (chunk: Uint8Array) => {
-        await this.binary.writeBytes(remotePath, this.uploadedBytes, false, chunk)
+        const result = await this.projectFiles.writeBytes(
+          remotePath,
+          this.uploadedBytes,
+          false,
+          chunk,
+        )
+        if (!result.ok) throw result.error
         this.checksum.update(chunk)
         this.uploadedBytes += BigInt(chunk.length)
         const bytes = Number(this.uploadedBytes)
@@ -127,13 +140,13 @@ export class Uploader {
       close: cleanup,
       abort: async (reason: string) => {
         cleanup()
-        await this.rpc.deleteFile(remotePath)
+        await this.projectFiles.deleteFile(remotePath)
         throw new Error(`Uploading process aborted. ${reason}`)
       },
     })
     // Disabled until https://github.com/enso-org/enso/issues/6691 is fixed.
     // Plus, handle the error here, as it should be displayed to the user.
-    // uploader.assertChecksum(remotePath)
+    // this.projectFiles.assertChecksum(remotePath)
     await this.file.stream().pipeTo(writableStream)
     return Ok({ source: 'Project', name: name.value })
   }
@@ -141,76 +154,4 @@ export class Uploader {
   private cleanup(name: string) {
     this.awareness.removeUpload(name)
   }
-
-  private async assertChecksum(path: Path): Promise<Result<void>> {
-    const engineChecksum = await this.rpc.fileChecksum(path)
-    if (!engineChecksum.ok) return engineChecksum
-    const hexChecksum = bytesToHex(this.checksum.digest())
-    if (hexChecksum != engineChecksum.value.checksum) {
-      return Err(
-        `Uploading file failed, checksum does not match. ${hexChecksum} != ${engineChecksum.value.checksum}`,
-      )
-    } else {
-      return Ok()
-    }
-  }
-
-  private dataDirPath(): Path {
-    return { rootId: this.projectRootId, segments: [DATA_DIR_NAME] }
-  }
-
-  private async ensureDataDirExists() {
-    const exists = await this.dataDirExists()
-    if (!exists.ok) return exists
-    if (exists.value) return Ok()
-    return await withContext(
-      () => 'When creating directory for uploaded file',
-      async () => {
-        return await this.rpc.createFile({
-          type: 'Directory',
-          name: DATA_DIR_NAME,
-          path: { rootId: this.projectRootId, segments: [] },
-        })
-      },
-    )
-  }
-
-  private async dataDirExists(): Promise<Result<boolean>> {
-    const info = await this.rpc.fileInfo(this.dataDirPath())
-    if (info.ok) return Ok(info.value.attributes.kind.type == 'Directory')
-    else if (
-      info.error.payload.cause instanceof RemoteRpcError &&
-      (info.error.payload.cause.code === ErrorCode.FILE_NOT_FOUND ||
-        info.error.payload.cause.code === ErrorCode.CONTENT_ROOT_NOT_FOUND)
-    ) {
-      return Ok(false)
-    } else {
-      return info
-    }
-  }
-
-  private async pickUniqueName(suggestedName: string): Promise<Result<string>> {
-    const files = await this.rpc.listFiles(this.dataDirPath())
-    if (!files.ok) return files
-    const existingNames = new Set(files.value.paths.map((path) => path.name))
-    const { stem, extension = '' } = splitFilename(suggestedName)
-    let candidate = suggestedName
-    let num = 1
-    while (existingNames.has(candidate)) {
-      candidate = `${stem}_${num}.${extension}`
-      num += 1
-    }
-    return Ok(candidate)
-  }
-}
-
-/** Split filename into stem and (optional) extension. */
-function splitFilename(fileName: string): { stem: string; extension?: string } {
-  const dotIndex = fileName.lastIndexOf('.')
-  if (dotIndex !== -1 && dotIndex !== 0) {
-    const stem = fileName.substring(0, dotIndex)
-    const extension = fileName.substring(dotIndex + 1)
-    return { stem, extension }
-  }
-  return { stem: fileName }
 }